# ldap-sync/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for synchronizing LDAP entries between two servers.",
        "title": "ldap-sync API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5500",
    "basePath": "/",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Returns OK if the application is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "probes"
                ],
                "summary": "Liveness Probe",
                "responses": {
                    "200": {
                        "description": "status: ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hook": {
            "post": {
                "description": "Receives an LDAP entry, applies transformation logic, and returns the transformed entry along with derived search specifications.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Process LDAP hook payload",
                "parameters": [
                    {
                        "description": "LDAP Entry Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.HookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.HookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Returns OK if the application is ready to serve traffic.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "probes"
                ],
                "summary": "Readiness Probe",
                "responses": {
                    "200": {
                        "description": "status: ready",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/results/{id}": {
            "get": {
                "description": "Retrieves all LDAP objects for a given search id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "results"
                ],
                "summary": "Get search results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique search id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Return full result (DN and content) if true, else only DN",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When full is true",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ResultEntryFull"
                            }
                        }
                    },
                    "404": {
                        "description": "Search results not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Retrieves a specific search by id if provided, or all searches if no id is specified.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Get search(s)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When id is provided\" or {array} SearchInfo \"When id is not provided",
                        "schema": {
                            "$ref": "#/definitions/main.SearchInfo"
                        }
                    },
                    "404": {
                        "description": "Search not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new search with a unique id. Returns an error if the id already exists.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Create new search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique search id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "LDAP search filter",
                        "name": "filter",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Refresh interval in seconds",
                        "name": "refresh",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional base DN for the search; defaults to global config if omitted",
                        "name": "baseDN",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters or search already exists",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/search/{id}": {
            "put": {
                "description": "Updates an existing search (complete replacement) with new filter, refresh, and optionally baseDN. If baseDN is omitted, the global config's BaseDN is used.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Update existing search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique search id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "LDAP search filter",
                        "name": "filter",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Refresh interval in seconds",
                        "name": "refresh",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional base DN for the search; defaults to global config if omitted",
                        "name": "baseDN",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters or search does not exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing search by its unique id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Delete search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique search id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Search not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.DerivedSearchSpec": {
            "type": "object",
            "properties": {
                "baseDN": {
                    "type": "string"
                },
                "filter": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "refresh": {
                    "type": "integer"
                }
            }
        },
        "main.HookRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "dn": {
                    "type": "string"
                }
            }
        },
        "main.HookResponse": {
            "type": "object",
            "properties": {
                "derived": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.DerivedSearchSpec"
                    }
                },
                "transformed": {
                    "$ref": "#/definitions/main.TransformedEntry"
                }
            }
        },
        "main.ResultEntryFull": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "dn": {
                    "type": "string"
                }
            }
        },
        "main.ResultEntrySimple": {
            "type": "object",
            "properties": {
                "dn": {
                    "type": "string"
                }
            }
        },
        "main.SearchInfo": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "refresh": {
                    "type": "integer"
                }
            }
        },
        "main.TransformedEntry": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "dn": {
                    "type": "string"
                }
            }
        }
    }
}